backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic==2.7.1
PyPDF2==3.0.1
langchain==0.1.20
langchain-openai==0.1.6
faiss-cpu==1.8.0